requirements.txt
langchain-groq
sqlalchemy  
langchain 
psycopg2-binary
flask
python-dotenv